Move db related logic to database.go
